temp_requirements.txt
torch>=2.0.0
transformers>=4.30.0
mistralai>=1.7.0
pytesseract>=0.3.10
pillow>=9.5.0
opencv-python>=4.7.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.17.0
jinja2>=3.1.2
pdf2image>=1.16.3
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pydantic>=2.0.0 